requirements.txt
Flask==2.3.2
gunicorn==20.1.0  # Needed for deploying to platforms like Render/Heroku
Flask
flask-cors
python-dotenv

